+++ baseline/app_version.py
@@ -1,4 +1,4 @@
 import os
 
 
-APP_VERSION = os.getenv("FAKE_UI_VERSION", "2.1.1")
+APP_VERSION = os.getenv("FAKE_UI_VERSION", "2.1.2")